Fix narrowing issue

+++ src/GL_setup.cpp
@@ -238,7 +238,7 @@ void setup_freetype(Param &par){
             texture,
             glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
             glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-            face->glyph->advance.x
+            (GLuint) face->glyph->advance.x
         };
         par.chmap.insert(std::pair<GLchar, Character>(c, character));
     }
